==> RiskBusters_Codes/1_data_cleaning_develop.ipynb <==
import pandas as pd
from clean_df import clean_df

df = pd.read_csv('../data/input/in_time.csv')

## Data cleaning - train

We clean the traing dataset and in parallel develop a function clean_df due to the high size of the file so that every team member can clean the data independently without uploading the whole cleeaning dataframe.

df

df_sum = df.isna().sum()
df_sum[df_sum > 0]

We do not observe any NAs, but we observe that NAs are shown as -9999

# change all -9999 to NaN
df = df.replace(-9999, pd.np.nan)

# show values and columns where number of NANs is greater than 0 
df_sum1 = df.isnull().sum()
df_sum1[df_sum1 > 0]

## Handle NAs

# change NAs in Active_mortgages to 0
df['Active_mortgages'] = df['Active_mortgages'].fillna(0)

# change NAs in selected columns to 0
columns_to_fill_with_0 = ['Active_mortgages', 
                          'Active_credit_card_lines', 
                          'External_term_loan_balance',
                          'External_mortgage_balance',
                          'External_credit_card_balance']

for column in columns_to_fill_with_0:
    df[column] = df[column].fillna(0)

# change NAs in all columns which contains selected names in column name to 0
columns_to_fill_with_0 = ['limit_in_revolving_loans_', 'utilized_limit_in_revolving_loans_']

for column in df.columns:
    for name in columns_to_fill_with_0:
        if name in column:
            df[column] = df[column].fillna(0)

df_sum2 = df.isnull().sum()
df_sum2[df_sum2 > 0]

## Data cleaning - test

We need to ensure that the testing dataset has the same problems as training dataset so we check it also here using the dunction clean_df developed basing on training dataframe.

df_test = pd.read_csv('../data/input/out_of_time.csv')

df_sum = df_test.isna().sum()
df_sum[df_sum > 0]

We also do not have any NAs, lets replace -9999 with NAa

# change all -9999 to NaN
df_test = df_test.replace(-9999, pd.np.nan)

df_sum = df_test.isna().sum()
df_sum[df_sum > 0]

We have exactly the same issues on this dataset, we can use the same funcion.

df_test = clean_df(df_test)
