==> RiskBusters_Codes/2_data_cleaning_reuse.ipynb <==
File to clean the data

import pandas as pd
from clean_df import clean_df

in_time = pd.read_csv('../data/input/in_time.csv')
out_of_time = pd.read_csv('../data/input/out_of_time.csv')

in_time = clean_df(in_time)
out_of_time = clean_df(out_of_time)

in_time.to_csv('../data/output/in_time_cleaned.csv', index=False)
out_of_time.to_csv('../data/output/out_of_time_cleaned.csv', index=False)
