==> RiskBusters_Codes/3_feature_engeneering.ipynb <==
import pandas as pd
import numpy as np
from feature_engeneering import modify_df
from feature_engeneering import drop_columns
from feature_engeneering import create_differences_columns

df = pd.read_csv('../data/output/in_time_cleaned.csv')

df_modified = modify_df(df)

columns = df_modified.columns
columns

# write colnames of df_modified to dataframe
df_colnames = pd.DataFrame(columns)
df_colnames

## delete columns

columns_to_delete_H = ['inc_transactions_amt_H', 'out_transactions_amt_H', 'overdue_term_loan_H', 
                       'overdue_credit_card_H', 'overdue_mortgage_H', 'Income_H',
                       'current_amount_balance_H', 'savings_amount_balance_H', 'inc_transactions_amt_H',
                        'Savings_amount_balance_col_H', 'Current_amount_balance_col_H',
                        'Overdue_term_loan_H', 'Overdue_credit_card_H', 'Overdue_mortgage_H',
                        'inc_transactions_H', 'out_transactions_H', 'inc_transactions_amt_H', 'out_transactions_amt_H',
                        'utilized_limit_in_revolving_loans_H', 'limit_in_revolving_loans_H']

columns_to_delete = ['Current_installment', 'Birth_date', 'Ref_month',
                     'Contract_end_date', 'Oldest_account_date']

df_tes_function = df_modified.copy()

# list of columns which has _H in the name in df_tes 

columns_H = [col for col in df_tes_function.columns if '_H0' in col]

# delete last letter from oevery column in columns_H
columns_H = [col[:-1] for col in columns_H]

columns_H

df_tes_function = create_differences_columns(df_tes_function, columns_H)

# drop columns with _H in the name

def drop_columns_after_transformation(df):
    for colname in columns_H:
        for i in range(13):
                colname2 = colname + str(i)
                if colname2 in df.columns:
                    df = df.drop(colname2, axis=1)
    return df

df_tes_function = drop_columns_after_transformation(df_tes_function)

columns = df_tes_function.columns
columns = pd.DataFrame(columns)
columns

nulls = df_tes_function.isna().sum()
nulls[nulls > 0]

# drop all NaN values
df_tes_function = df_tes_function.dropna()

## clients_age

# change the type of Birth_date to datetime
df['Birth_date'] = pd.to_datetime(df['Birth_date'])

# add a column clients_age with the age of the client by substracting the birth_date from the current date
current_date = pd.to_datetime('2024-04-03')
df['clients_age'] = (current_date - df['Birth_date']).astype('<m8[Y]').astype(int)

## Create difference columns

Approach to new variables:

1. We crate 12 new variables for every Hx
2. We calculate for every column:

(H0 - H3) / H0
(H0 - H6) / H0
(H0 - H12) / H0

## Create DTI

(inc_transactions_amt - out_transactions_amt) / current_installment

columns_to_delete_H = ['inc_transactions_amt_H', 'out_transactions_amt_H']
columns_to_delete = ['Current_installment', 'Birth_date']
