==> RiskBusters_Codes/4.feature_seleciton.ipynb <==
from sklearn.linear_model import ElasticNetCV
from sklearn.preprocessing import StandardScaler
from sklearn.feature_selection import SelectFromModel
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.feature_selection import mutual_info_regression
from sklearn.feature_selection import VarianceThreshold
from sklearn.feature_selection import f_regression
from sklearn.linear_model import ElasticNet
from sklearn.model_selection import cross_val_score
from itertools import product
from sklearn.ensemble import RandomForestRegressor
import pandas as pd 
import numpy as np 

df_train = pd.read_csv('..\\data\\output\\fe\\in_time_fe.csv', sep=',')
#df_test =  pd.read_csv('..\data\output\fe\out_of_time_cleaned.csv', sep=',')

X_col = df_train.drop(columns=['Target']).columns.tolist()

y_col = 'Target'

Mi-value

mi = dict()
for i in X_col:
    mi.update({i: mutual_info_regression(df_train[[i]].values, df_train[y_col].values)[0]})

miDF = pd.DataFrame.from_dict(mi, orient="index", columns=["score"])

miDF.sort_values(by="score", ascending = False).head(9)

F-test

import warnings

warnings.simplefilter(action="ignore", category=FutureWarning)
fscore = dict()
for i in X_col:
    fscore.update({i: f_regression(df_train[[i]].values, df_train[y_col].values)[1]})

fscoreDF = pd.DataFrame.from_dict(fscore, orient="index", columns=["p_value_score"])

fscoreDF.sort_values(by="p_value_score").head(9)

Corr

corr = list()
for i in x_col:
    corr.append(stats.spearmanr(df.loc[:, y_col].values, df.loc[:, i].values)[0])

df_train=df_train.drop(['Customer_id', "Contract_origination_date",'Ref_month_dt', 'Oldest_account_date_dt','Contract_end_date_dt','Unnamed: 0'], axis=1)

X = df_train.drop(columns = ['Target'])
y = df_train['Target']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, stratify=y, random_state=42)


scaler = StandardScaler()
X_scaled = scaler.fit_transform(X)

elastic_net = ElasticNetCV(cv=3, random_state=42)

elastic_net.fit(X_scaled, y)

selected_features = SelectFromModel(elastic_net, prefit=True, max_features=9)
X_selected = selected_features.transform(X_scaled)

selected_indices = selected_features.get_support(indices=True)

selected_feature_names = X.columns[selected_indices]


print("Selected Features:")
print(selected_feature_names)


X_train = df_train[X_col]
y_train = df_train[y_col]

rf_model = RandomForestClassifier(random_state = 420, n_jobs = -1)

fi = rf_model.fit(X, y).feature_importances_
fi_9 = pd.DataFrame(fi, index = X_col, columns = ["Feature Importance"]).sort_values(by = "Feature Importance", ascending = False).head(9)
print(fi_9)

variables_rf=fi_9.index

#df_train[variables_rf]
df_model_train=pd.concat([df_train[variables_rf], y], axis=1)

df_model_train

Shapley values

pipeline.fit(X_train, y_train);
model = pipeline['clf']

explainer = shap.Explainer(model, 
                           pipeline['prep'].transform(X_train), 
                           feature_names=X_train.columns)
shap_values = explainer(pipeline['prep'].transform(X_train))

fig = shap.plots.waterfall(shap_values[1])
plt.tight_layout()

PSI Comparison

models = {
    #'Random Forest': RandomForestClassifier(random_state=123),
    #'XGBoost': XGBClassifier(random_state=123, n_jobs=-1, objective='binary:logistic'),
    'LightGBM': LGBMClassifier(random_state=123),
    #'SVC': SVC(probability=True, random_state=123),
    #'CatBoost': CatBoostClassifier(random_state=123, verbose=0),
    #'KNN': KNeighborsClassifier(),
    'Logistic Regression': LogisticRegression(random_state=123, max_iter = 1000),
    'Decision Tree': DecisionTreeClassifier(random_state=123)
}

# Define parameter grids for each model
#param_grid_rf = {
#    'n_estimators': [100, 200, 300],
#    'max_depth': [4, 6, 8],
#    'min_samples_split': [2, 5, 10]
#}

param_grid_lgbm = {
    'learning_rate': [0.01, 0.1, 0.5],
    'n_estimators': [60, 80, 100],
    'max_depth': [4, 6, 8],
    'subsample': [0.8, 1.0],
    'colsample_bytree': [0.8, 1.0]
}

param_grid_logreg = {
    'C': [ 0.01, 0.1, 1.0],
    'penalty': ['l1', 'l2'],
    'solver': ['liblinear', 'saga']
}

param_grid_dt = {
    'model__max_depth': [3, 5, 7],
    #'model__min_samples_split': [2, 5, 10],
    'model__min_samples_leaf': [10, 20, 30],
    'model__criterion': ['gini', 'entropy']
}

X = df_train.drop(columns = ['Target'])
y = df_train['Target']
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, stratify=y, random_state=42)


rus = RandomUnderSampler(random_state=123)
scaler = RobustScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

# Apply Random Under-Sampling
X_resampled, y_resampled = rus.fit_resample(X_train_scaled, y_train)




X_t = df_test.drop(columns = ['Target'])
y_t = df_test['Target']

X_t_scaled = scaler.transform(X_t)

y_pred = best_model.predict(X_t_scaled)
print(classification_report(y_t, y_pred))

cm = confusion_matrix(y_t, y_pred)
sns.heatmap(cm, annot=True, fmt='d', cmap='Blues', cbar=False)
plt.xlabel('Predicted labels')
plt.ylabel('True labels')
plt.title(f'Confusion Matrix - {name}\nAUC-ROC Score: {auc_roc:.3f}')

plt.plot(fpr, tpr, label=f'{name} (AUC = {auc_roc:.3f})')
plt.plot([0, 1], [0, 1], linestyle='--', lw=2, color='r', label='Chance', alpha=0.8)
plt.xlabel('False Positive Rate')
plt.ylabel('True Positive Rate')
plt.title(f'AUC-ROC Curve - {name}')
plt.legend()



def calculate_psi(expected, actual, buckets=10):
    """Calculate Population Stability Index (PSI) for two distributions."""
    
    def calculate_bucketed_counts(data, buckets):
        # Calculate bucket boundaries
        quantiles = np.linspace(0, 100, num=buckets + 1)
        bin_edges = np.percentile(data, q=quantiles)
        
        # Assign data points to buckets
        bucket_indices = np.digitize(data, bin_edges[1:-1])
        
        # Calculate counts for each bucket
        bucket_counts = pd.Series(bucket_indices).value_counts(sort=False)
        
        return bin_edges, bucket_counts
    
    def calculate_psi_bucket(expected_counts, actual_counts):
        expected_freq = expected_counts / expected_counts.sum()
        actual_freq = actual_counts / actual_counts.sum()
        
        # Calculate PSI for each bucket
        psi_bucket = (actual_freq - expected_freq) * np.log(actual_freq / expected_freq)
        
        return psi_bucket.sum()
    
    # Calculate PSI for each variable
    psi_values = {}
    for col in expected.columns:
        expected_data = expected[col].dropna()
        actual_data = actual[col].dropna()
        
        # Calculate bucketed counts for expected and actual data
        expected_bin_edges, expected_counts = calculate_bucketed_counts(expected_data, buckets)
        actual_bin_edges, actual_counts = calculate_bucketed_counts(actual_data, buckets)
        
        # Extend bin edges for consistent calculations
        all_bin_edges = np.unique(np.concatenate([expected_bin_edges, actual_bin_edges]))
        
        # Re-calculate bucketed counts with consistent bin edges
        expected_counts = pd.cut(expected_data, bins=all_bin_edges, include_lowest=True).value_counts(sort=False)
        actual_counts = pd.cut(actual_data, bins=all_bin_edges, include_lowest=True).value_counts(sort=False)
        
        # Calculate PSI for the variable
        psi = calculate_psi_bucket(expected_counts, actual_counts)
        
        psi_values[col] = psi
    
    return psi_values

# Calculate PSI for each variable between df_1 and df_2
psi_values = calculate_psi(df_train[x_col], df_test[x_col])

# Visualize PSI values
psi_df = pd.DataFrame.from_dict(psi_values, orient='index', columns=['PSI'])
psi_df = psi_df.sort_values(by='PSI', ascending=False)

# Plot the PSI values
plt.figure(figsize=(12, 6))
sns.barplot(x=psi_df.index, y='PSI', data=psi_df, palette='viridis')
plt.xticks(rotation=90)
plt.xlabel('Variable')
plt.ylabel('Population Stability Index (PSI)')
plt.title('Population Stability Index (PSI) for Variables')
plt.tight_layout()
plt.show()
