==> RiskBusters_Codes/4_feature_eng_reuse.ipynb <==
import pandas as pd
from feature_engeneering import modify_df
from feature_engeneering import drop_columns
from feature_engeneering import create_differences_columns
from feature_engeneering import drop_columns_after_transformation

df = pd.read_csv('../data/output/in_time_cleaned.csv')

df = modify_df(df)

columns_to_delete_H = ['inc_transactions_amt_H', 'out_transactions_amt_H',
                       'overdue_credit_card_H', 'overdue_mortgage_H', 'income_H',
                       'current_amount_balance_H', 'savings_amount_balance_H', 'inc_transactions_amt_H',
                        'Savings_amount_balance_col_H', 'Current_amount_balance_col_H',
                        'Overdue_term_loan_H', 'Overdue_credit_card_H', 'Overdue_mortgage_H',
                        'inc_transactions_H', 'out_transactions_H', 'inc_transactions_amt_H', 'out_transactions_amt_H',
                        'utilized_limit_in_revolving_loans_H', 'limit_in_revolving_loans_H']

columns_to_delete = ['Current_installment', 'Birth_date', 'Ref_month',
                     'Contract_end_date', 'Oldest_account_date']

# Generate a list of columns to drop
columns_to_drop = [col + str(i) for col in columns_to_delete_H for i in range(1, 13)]

# Filter out columns that exist in the DataFrame
columns_to_drop = [col for col in columns_to_drop if col in df.columns]
columns_to_drop

df = drop_columns(df, columns_to_delete_H, columns_to_delete)

# list of columns which has _H in the name in df_tes 
columns_H = [col for col in df.columns if '_H12' in col]

# delete last letter from oevery column in columns_H
columns_H = [col[:-2] for col in columns_H]
columns_H

df = create_differences_columns(df, columns_H)

columns = df.columns
columns = pd.DataFrame(columns)
columns

df = drop_columns_after_transformation(df, columns_H)

columns = df.columns
columns = pd.DataFrame(columns)
columns

nulls = df.isna().sum()
nulls[nulls > 0]

# drop all NaN values
df = df.dropna()

df.to_csv('../data/output/fe/in_time_fe.csv')

## DF test

df = pd.read_csv('../data/output/out_of_time_cleaned.csv')

df = modify_df(df)
df = drop_columns(df, columns_to_delete_H, columns_to_delete)
df = create_differences_columns(df, columns_H)

df = drop_columns_after_transformation(df, columns_H)

# drop all NaN values
df = df.dropna()

df.to_csv('../data/output/fe/out_of_time_fe.csv')
